Избавляется от ненужных модулей
