+++ js/form-listener-submit.js
@@ -1,7 +1,7 @@
-import { form } from './global-constant.js';
-import { validateForm } from './validate-form.js';
+// import { form } from './global-constant.js';
+// import { validateForm } from './validate-form.js';
 
-const removeListenerModal = () => form.removeEventListener('submit', validateForm);
-const addListenerModal = () => form.addEventListener('submit', validateForm);
+// const removeListenerModal = () => form.removeEventListener('submit', validateForm);
+// const addListenerModal = () => form.addEventListener('submit', validateForm);
 
-export { removeListenerModal, addListenerModal };
+// export { removeListenerModal, addListenerModal };
