+++ js/get-photo-gallery-list.js
@@ -1,0 +1,10 @@
+import { getPhotoData } from './api.js';
+
+const getPhotoGalleryList = function () {
+  return getPhotoData()
+    .then((photos) => photos);
+};
+
+const dataPhotoGallery = await getPhotoGalleryList();
+
+export { dataPhotoGallery };
