+++ js/main.js
@@ -1,23 +1,6 @@
 import './generate-photo-gallery/generate-photo-gallery.js';
 import './rendering-pictures.js';
 import './big-picture.js';
-import './user-form.js'; // Почему работает без импорта?
 import './validate-form.js';
 import './editing-picture.js';
-// import { getData, sendData } from './api.js';
 
-
-// const setHui = async function (data) {
-//   try {
-//     await sendData(data);
-//   } catch {
-//     showErrorMessage();
-//   }
-// };
-
-// try {
-//   const data = await getData();
-//   renderGallery(data);
-// } catch (err) {
-//   showAllert(err.message);
-// }
